move spmerging model definition outside of the specify app and into notifications to prevent errors
